refactor: refine plugin configuration types
- Update plugin exports across all packages to use refined types
- Move CompatibleConfig and CompatiblePlugin interfaces to kit package
- Update configuration files to use new type structure
- Remove outdated documentation for compatibility interfaces
