+++ packages/plugins/eslint-plugin-react-debug/src/configs/all.ts
@@ -1,4 +1,4 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 import { DEFAULT_ESLINT_REACT_SETTINGS } from "@eslint-react/shared";
 
 export const name = "react-debug/all";
@@ -9,7 +9,7 @@ export const rules = {
   "react-debug/hook": "warn",
   "react-debug/is-from-react": "warn",
   "react-debug/jsx": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
 
 export const settings = {
   "react-x": DEFAULT_ESLINT_REACT_SETTINGS,
