+++ packages/plugins/eslint-plugin-react-hooks-extra/src/configs/recommended.ts
@@ -1,7 +1,7 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 export const name = "react-hooks-extra/recommended";
 
 export const rules = {
   "react-hooks-extra/no-direct-set-state-in-use-effect": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
