+++ packages/plugins/eslint-plugin-react-hooks-extra/src/index.ts
@@ -1,5 +1,5 @@
-import type { RulePreset } from "@eslint-react/kit";
-import type { CompatibleConfig } from "@eslint-react/shared";
+import type { RuleConfig } from "@eslint-react/kit";
+import type { CompatibleConfig } from "@eslint-react/kit";
 
 import * as recommendedConfig from "./configs/recommended";
 import { plugin } from "./plugin";
@@ -13,7 +13,7 @@ function toFlatConfig(config: CompatibleConfig) {
   };
 }
 
-function toLegacyConfig({ rules }: { rules: RulePreset }) {
+function toLegacyConfig({ rules }: CompatibleConfig) {
   return {
     plugins: ["react-hooks-extra"],
     rules,
