+++ packages/plugins/eslint-plugin-react-naming-convention/src/configs/recommended.ts
@@ -1,8 +1,8 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 export const name = "react-naming-convention/recommended";
 
 export const rules = {
   "react-naming-convention/context-name": "warn",
   // "react-naming-convention/use-state": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
