+++ packages/plugins/eslint-plugin-react-naming-convention/src/plugin.ts
@@ -1,4 +1,4 @@
-import type { CompatiblePlugin } from "@eslint-react/shared";
+import type { CompatiblePlugin } from "@eslint-react/kit";
 
 import { name, version } from "../package.json";
