+++ packages/plugins/eslint-plugin-react-web-api/src/configs/recommended.ts
@@ -1,4 +1,4 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 import { DEFAULT_ESLINT_REACT_SETTINGS } from "@eslint-react/shared";
 
 export const name = "react-web-api/recommended";
@@ -8,7 +8,7 @@ export const rules = {
   "react-web-api/no-leaked-interval": "warn",
   "react-web-api/no-leaked-resize-observer": "warn",
   "react-web-api/no-leaked-timeout": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
 
 export const settings = {
   "react-x": DEFAULT_ESLINT_REACT_SETTINGS,
