+++ packages/plugins/eslint-plugin-react-x/src/configs/recommended-type-checked.ts
@@ -1,4 +1,4 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 import * as recommendedTypeScript from "./recommended-typescript";
 
@@ -8,7 +8,7 @@ export const rules = {
   ...recommendedTypeScript.rules,
   "react-x/no-leaked-conditional-rendering": "warn",
   "react-x/no-unused-props": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
 
 export const settings = {
   ...recommendedTypeScript.settings,
