+++ packages/plugins/eslint-plugin/src/configs/all.ts
@@ -1,4 +1,4 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 import reactDebug from "eslint-plugin-react-debug";
 import reactHooksExtra from "eslint-plugin-react-hooks-extra";
 import reactNamingConvention from "eslint-plugin-react-naming-convention";
@@ -104,7 +104,7 @@ export const rules = {
   "@eslint-react/naming-convention/filename": "warn",
   "@eslint-react/naming-convention/filename-extension": "warn",
   "@eslint-react/naming-convention/use-state": "warn",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
 
 export const plugins = {
   ...x.plugins,
