+++ packages/plugins/eslint-plugin/src/configs/debug.ts
@@ -1,3 +1,5 @@
+import type { RuleConfig } from "@eslint-react/kit";
+
 import reactDebug from "eslint-plugin-react-debug";
 
 export const name = "@eslint-react/debug";
@@ -8,7 +10,7 @@ export const rules = {
   "@eslint-react/debug/hook": "warn",
   "@eslint-react/debug/is-from-react": "warn",
   "@eslint-react/debug/jsx": "warn",
-} as const;
+} as const satisfies Record<string, RuleConfig>;
 
 export const plugins = {
   "@eslint-react/debug": reactDebug,
