+++ packages/plugins/eslint-plugin/src/configs/disable-debug.ts
@@ -1,7 +1,9 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 import { rules as debugRules } from "./debug";
 
 export const name = "@eslint-react/disable-debug";
 
-export const rules: RulePreset = Object.fromEntries(Object.entries(debugRules).map(([rule]) => [rule, "off"]));
+export const rules: Record<string, RuleConfig> = Object.fromEntries(
+  Object.entries(debugRules).map(([rule]) => [rule, "off"]),
+);
