+++ packages/plugins/eslint-plugin/src/configs/disable-dom.ts
@@ -1,7 +1,9 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 import { rules as domRules } from "./dom";
 
 export const name = "@eslint-react/disable-dom";
 
-export const rules: RulePreset = Object.fromEntries(Object.entries(domRules).map(([key]) => [key, "off"] as const));
+export const rules: Record<string, RuleConfig> = Object.fromEntries(
+  Object.entries(domRules).map(([key]) => [key, "off"] as const),
+);
