+++ packages/plugins/eslint-plugin/src/configs/disable-type-checked.ts
@@ -1,8 +1,8 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 export const name = "@eslint-react/disable-type-checked";
 
-export const rules: RulePreset = {
+export const rules: Record<string, RuleConfig> = {
   "@eslint-react/no-leaked-conditional-rendering": "off",
   "@eslint-react/no-unused-props": "off",
   "@eslint-react/prefer-read-only-props": "off",
