+++ packages/plugins/eslint-plugin/src/configs/disable-web-api.ts
@@ -1,7 +1,9 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 
 import { rules as webApiRules } from "./web-api";
 
 export const name = "@eslint-react/disable-web-api";
 
-export const rules: RulePreset = Object.fromEntries(Object.entries(webApiRules).map(([key]) => [key, "off"] as const));
+export const rules: Record<string, RuleConfig> = Object.fromEntries(
+  Object.entries(webApiRules).map(([key]) => [key, "off"] as const),
+);
