+++ packages/plugins/eslint-plugin/src/configs/dom.ts
@@ -1,4 +1,4 @@
-import type { RulePreset } from "@eslint-react/kit";
+import type { RuleConfig } from "@eslint-react/kit";
 import { DEFAULT_ESLINT_REACT_SETTINGS } from "@eslint-react/shared";
 import reactDom from "eslint-plugin-react-dom";
 
@@ -20,7 +20,7 @@ export const rules = {
   "@eslint-react/dom/no-unsafe-target-blank": "warn",
   "@eslint-react/dom/no-use-form-state": "error",
   "@eslint-react/dom/no-void-elements-with-children": "error",
-} as const satisfies RulePreset;
+} as const satisfies Record<string, RuleConfig>;
 
 export const plugins = {
   "@eslint-react/dom": reactDom,
