+++ packages/shared/docs/README.md
@@ -6,8 +6,6 @@
 
 ## Interfaces
 
-- [CompatibleConfig](interfaces/CompatibleConfig.md)
-- [CompatiblePlugin](interfaces/CompatiblePlugin.md)
 - [CustomComponentNormalized](interfaces/CustomComponentNormalized.md)
 - [CustomComponentPropNormalized](interfaces/CustomComponentPropNormalized.md)
 - [ESLintReactSettingsNormalized](interfaces/ESLintReactSettingsNormalized.md)
