--- packages/shared/docs/interfaces/CompatibleConfig.md
@@ -1,19 +1,0 @@
-[**@eslint-react/shared**](../README.md)
-
-***
-
-[@eslint-react/shared](../README.md) / CompatibleConfig
-
-# Interface: CompatibleConfig
-
-## Properties
-
-### name?
-
-> `optional` **name**: `string`
-
-***
-
-### rules?
-
-> `optional` **rules**: [`Record`](https://www.typescriptlang.org/docs/handbook/utility-types.html#recordkeys-type)\<`string`, `any`\>
