--- packages/shared/docs/interfaces/CompatiblePlugin.md
@@ -1,27 +1,0 @@
-[**@eslint-react/shared**](../README.md)
-
-***
-
-[@eslint-react/shared](../README.md) / CompatiblePlugin
-
-# Interface: CompatiblePlugin
-
-## Properties
-
-### meta
-
-> **meta**: `object`
-
-#### name
-
-> **name**: `string`
-
-#### version
-
-> **version**: `string`
-
-***
-
-### rules
-
-> **rules**: [`Record`](https://www.typescriptlang.org/docs/handbook/utility-types.html#recordkeys-type)\<`string`, `any`\>
