+++ packages/shared/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./_id";
 export * from "./_require";
-export type * from "./compatibility-types";
 export * from "./constants";
 export * from "./get-doc-url";
 export * from "./get-react-version";
